Bundled the Cascade Classifiers with the project.
Bundled the Cascade Classifiers with the project and updated the project
with the new file paths. There shouldn't be any problems moving this to
other systems as long as they have the  Processing and OpenCV libraries.
Added functionality that exits the application if it can't open a camera
stream; as well as if it can't find the CascadeClassifiers.