Image print-outs and exit actual
Processing should now completely terminate the application when the user
clicks to close the window. Another feature that has been added is the
at each stage the application writes an image of what's been captured to
several folders; although since I'm not cropping the search area
properly it's not working.